Folders structure was changed in frontend/src/main/web/src/app. New folder "pages" was created. Landing and not-found-page were moved to "pages" folder. All the components of landing page (carousel, features and so on) were moved to "/pages/landing". Test component was deleted as far as we don't need to make tests for frontend
