+++ frontend/src/main/web/src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { LandingComponent } from './landing/landing.component';
-import { NotFoundPageComponent } from './not-found-page/not-found-page.component';
+import { LandingComponent } from './pages/landing/landing.component';
+import { NotFoundPageComponent } from './pages/not-found-page/not-found-page.component';
 
 const routes: Routes = [
   { path: '', component: LandingComponent},
