+++ frontend/src/main/web/src/app/pages/landing/carousel/carousel.component.html
