+++ frontend/src/main/web/src/app/pages/landing/features/features.component.scss
