+++ frontend/src/main/web/src/app/pages/landing/footer/footer.component.scss
