+++ frontend/src/main/web/src/app/pages/landing/header/header.component.scss
