+++ frontend/src/main/web/src/app/pages/landing/landing.component.scss
