+++ frontend/src/main/web/src/app/pages/landing/map/map.component.ts
