+++ frontend/src/main/web/src/app/pages/landing/search-menu/search-menu.component.scss
