+++ frontend/src/main/web/src/app/pages/not-found-page/not-found-page.component.scss
