--- frontend/src/main/web/src/app/test/test.component.html
@@ -1,3 +1,0 @@
-<p>
-  test works!
-</p>
