--- frontend/src/main/web/src/app/test/test.component.scss
@@ -1,3 +1,0 @@
-p {
-  color: red;
-}
